Rename CreateFailureRetryDecider
